requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
click==8.1.3
    # via -r requirements.in
decorator==5.1.1
    # via ipython
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.13
    # via
    #   -r requirements.in
    #   django-phonenumber-field
django-phonenumber-field==6.1.0
    # via -r requirements.in
django-phonenumbers==1.0.1
    # via -r requirements.in
executing==0.8.3
    # via stack-data
gunicorn==20.1.0
    # via -r requirements.in
ipython==8.4.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
phonenumbers==8.12.50
    # via django-phonenumbers
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.29
    # via ipython
psycopg2==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.12.0
    # via ipython
pytz==2022.1
    # via django
six==1.16.0
    # via asttokens
sqlparse==0.4.2
    # via django
stack-data==0.2.0
    # via ipython
traitlets==5.2.2.post1
    # via
    #   ipython
    #   matplotlib-inline
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==6.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
